First Commit of Monotone Polygon Research
HEURISTIC SOLUTION FOR DISCRETE VERTEX GUARDING IN
MONOTONE POLYGONS | UNDERGRAD CAPSTONE
Feb 2013 – May 2013 | CCSC Midwest 2013
Worked with Erik Krohn to create Heuristic Solution for Discrete Vertex
Guarding in Monotone Polygons. My capstone project was related to the
Art Gallery Problem in the field of computational geometry. These
problems relate to polygon's visibility in the relation to the minimum
number of guards required to protect the polygon. I also presented a
poster on the same topic during the CCSC Midwest Conference at the
University of Findlay.